requirements.txt
azure-cognitiveservices-vision-computervision
pytest
pytest-asyncio
playwright
python-dotenv
